change to count the number of complete payment, pending payment

+++ service/payment-service/src/main/java/com/nls/paymentservice/api/dto/response/RevenueData.java
@@ -7,7 +7,7 @@ import java.math.BigDecimal;
 @Builder
 public record RevenueData(
         BigDecimal totalRevenue,
-        BigDecimal completedPayments,
-        BigDecimal pendingPayments
+        Long completedPayments,
+        Long pendingPayments
 ) {
 }
